refactor(transcription): extract shared OnnxAsrEngine base
Move the onnx-asr recognize() path and the timestamped-result mapper
into a new OnnxAsrEngine base class so multiple onnx-asr-backed engines
can share them. ParakeetEngine becomes a thin subclass parameterised by
class attributes (name, config_key, default_model_id, onnx_id_map).

map_parakeet_result is renamed map_onnx_asr_result (architecture-agnostic)
and re-exported from parakeet_engine under the old name for compatibility.
Parakeet behaviour is unchanged.
